requirements.txt
asgiref==3.5.0
beautifulsoup4==4.11.1
bootstrap4==0.1.0
cycler==0.10.0
Django==3.2.12
django-bootstrap-datepicker-plus==4.0.0
django-bootstrap4==22.1
importlib-metadata==2.1.3
kiwisolver==1.1.0
matplotlib==3.1.2
mysql-connector-python==8.0.18
numpy==1.17.4
Pillow==9.0.1
protobuf==3.11.1
pygame==1.9.6
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2021.3
req==1.0.0
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
typing-extensions==4.1.1
zipp==3.8.0
